[APM] Move from `/shared` folder where possible
