+++ x-pack/plugins/apm/public/components/app/dependency_operation_detail_view/index.tsx
@@ -17,7 +17,7 @@ import { useDependencyDetailOperationsBreadcrumb } from '../../../hooks/use_depe
 import { FETCH_STATUS, useFetcher } from '../../../hooks/use_fetcher';
 import { useTimeRange } from '../../../hooks/use_time_range';
 import { DependencyMetricCharts } from '../../shared/dependency_metric_charts';
-import { DetailViewHeader } from '../../shared/detail_view_header';
+import { DetailViewHeader } from './detail_view_header';
 import { ResettingHeightRetainer } from '../../shared/height_retainer/resetting_height_container';
 import { push, replace } from '../../shared/links/url_helpers';
 import { SortFunction } from '../../shared/managed_table';
