+++ x-pack/plugins/apm/public/components/routing/app_root/apm_header_action_menu/index.tsx
@@ -14,12 +14,12 @@ import {
 import { apmLabsButton } from '@kbn/observability-plugin/common';
 import { i18n } from '@kbn/i18n';
 import React from 'react';
-import { getAlertingCapabilities } from '../../alerting/utils/get_alerting_capabilities';
-import { getLegacyApmHref } from '../links/apm/apm_link';
-import { useApmPluginContext } from '../../../context/apm_plugin/use_apm_plugin_context';
+import { getAlertingCapabilities } from '../../../alerting/utils/get_alerting_capabilities';
+import { getLegacyApmHref } from '../../../shared/links/apm/apm_link';
+import { useApmPluginContext } from '../../../../context/apm_plugin/use_apm_plugin_context';
 import { AlertingPopoverAndFlyout } from './alerting_popover_flyout';
 import { AnomalyDetectionSetupLink } from './anomaly_detection_setup_link';
-import { useServiceName } from '../../../hooks/use_service_name';
+import { useServiceName } from '../../../../hooks/use_service_name';
 import { InspectorHeaderLink } from './inspector_header_link';
 import { Labs } from './labs';
