+++ x-pack/plugins/apm/public/components/routing/app_root/apm_header_action_menu/labs/labs_flyout.tsx
@@ -24,9 +24,9 @@ import {
 import { LazyField } from '@kbn/advanced-settings-plugin/public';
 import { i18n } from '@kbn/i18n';
 import React from 'react';
-import { useApmPluginContext } from '../../../../context/apm_plugin/use_apm_plugin_context';
-import { useApmEditableSettings } from '../../../../hooks/use_apm_editable_settings';
-import { useFetcher, isPending } from '../../../../hooks/use_fetcher';
+import { useApmPluginContext } from '../../../../../context/apm_plugin/use_apm_plugin_context';
+import { useApmEditableSettings } from '../../../../../hooks/use_apm_editable_settings';
+import { useFetcher, isPending } from '../../../../../hooks/use_fetcher';
 
 interface Props {
   onClose: () => void;
