+++ x-pack/plugins/apm/public/components/routing/app_root/redirect_dependencies_to_dependencies_inventory.tsx
