+++ x-pack/plugins/apm/public/components/routing/app_root/redirect_with_default_date_range/index.tsx
@@ -6,9 +6,9 @@
  */
 import { ReactElement } from 'react';
 import { useLocation } from 'react-router-dom';
-import { useApmRouter } from '../../../hooks/use_apm_router';
-import { useDateRangeRedirect } from '../../../hooks/use_date_range_redirect';
-import { isRouteWithTimeRange } from '../is_route_with_time_range';
+import { useApmRouter } from '../../../../hooks/use_apm_router';
+import { useDateRangeRedirect } from '../../../../hooks/use_date_range_redirect';
+import { isRouteWithTimeRange } from '../../../shared/is_route_with_time_range';
 
 // This is a top-level component that blocks rendering of the routes
 // if there is no valid date range, and redirects to one if needed.
