+++ x-pack/plugins/apm/public/components/routing/app_root/redirect_with_default_environment/index.test.tsx
@@ -10,9 +10,9 @@ import { render } from '@testing-library/react';
 import { createMemoryHistory, Location, MemoryHistory } from 'history';
 import qs from 'query-string';
 import { RedirectWithDefaultEnvironment } from '.';
-import { apmRouter } from '../../routing/apm_route_config';
-import * as useApmPluginContextExports from '../../../context/apm_plugin/use_apm_plugin_context';
-import { ENVIRONMENT_ALL } from '../../../../common/environment_filter_values';
+import { apmRouter } from '../../apm_route_config';
+import * as useApmPluginContextExports from '../../../../context/apm_plugin/use_apm_plugin_context';
+import { ENVIRONMENT_ALL } from '../../../../../common/environment_filter_values';
 
 describe('RedirectWithDefaultEnvironment', () => {
   let history: MemoryHistory;
