+++ x-pack/plugins/apm/public/components/routing/app_root/redirect_with_offset/index.test.tsx
@@ -10,9 +10,9 @@ import { render } from '@testing-library/react';
 import { createMemoryHistory, Location, MemoryHistory } from 'history';
 import qs from 'query-string';
 import { RedirectWithOffset } from '.';
-import { apmRouter } from '../../routing/apm_route_config';
-import * as useApmPluginContextExports from '../../../context/apm_plugin/use_apm_plugin_context';
-import { TimeRangeComparisonEnum } from '../time_comparison/get_comparison_options';
+import { apmRouter } from '../../apm_route_config';
+import * as useApmPluginContextExports from '../../../../context/apm_plugin/use_apm_plugin_context';
+import { TimeRangeComparisonEnum } from '../../../shared/time_comparison/get_comparison_options';
 
 describe('RedirectWithOffset', () => {
   let history: MemoryHistory;
