+++ x-pack/plugins/apm/public/components/routing/app_root/redirect_with_offset/index.tsx
@@ -8,16 +8,16 @@
 import { useLocation } from 'react-router-dom';
 import qs from 'query-string';
 import React from 'react';
-import { useApmRouter } from '../../../hooks/use_apm_router';
-import { isRouteWithComparison } from '../is_route_with_time_range';
+import { useApmRouter } from '../../../../hooks/use_apm_router';
+import { isRouteWithComparison } from '../../../shared/is_route_with_time_range';
 import {
   TimeRangeComparisonEnum,
   dayAndWeekBeforeToOffset,
-} from '../time_comparison/get_comparison_options';
-import { useApmPluginContext } from '../../../context/apm_plugin/use_apm_plugin_context';
-import { getComparisonEnabled } from '../time_comparison/get_comparison_enabled';
-import { toBoolean } from '../../../context/url_params_context/helpers';
-import { RenderRedirectTo } from '../../routing/redirect_to';
+} from '../../../shared/time_comparison/get_comparison_options';
+import { useApmPluginContext } from '../../../../context/apm_plugin/use_apm_plugin_context';
+import { getComparisonEnabled } from '../../../shared/time_comparison/get_comparison_enabled';
+import { toBoolean } from '../../../../context/url_params_context/helpers';
+import { RenderRedirectTo } from '../../redirect_to';
 
 export function RedirectWithOffset({
   children,
