+++ x-pack/plugins/apm/public/components/shared/links/dependency_link.tsx
@@ -9,10 +9,10 @@ import { EuiFlexGroup, EuiFlexItem, EuiLink } from '@elastic/eui';
 import { TypeOf } from '@kbn/typed-react-router-config';
 import React from 'react';
 import { euiStyled } from '@kbn/kibana-react-plugin/common';
-import { useApmRouter } from '../../hooks/use_apm_router';
-import { truncate } from '../../utils/style';
-import { ApmRoutes } from '../routing/apm_route_config';
-import { SpanIcon } from './span_icon';
+import { useApmRouter } from '../../../hooks/use_apm_router';
+import { truncate } from '../../../utils/style';
+import { ApmRoutes } from '../../routing/apm_route_config';
+import { SpanIcon } from '../span_icon';
 
 const StyledLink = euiStyled(EuiLink)`${truncate('100%')};`;
