+++ x-pack/plugins/apm/public/components/shared/search_bar/search_bar.tsx
@@ -13,13 +13,13 @@ import {
   EuiSpacer,
 } from '@elastic/eui';
 import React from 'react';
-import { isMobileAgentName } from '../../../common/agent_name';
-import { useApmServiceContext } from '../../context/apm_service/use_apm_service_context';
-import { useBreakpoints } from '../../hooks/use_breakpoints';
-import { ApmDatePicker } from './date_picker/apm_date_picker';
-import { KueryBar } from './kuery_bar';
-import { TimeComparison } from './time_comparison';
-import { TransactionTypeSelect } from './transaction_type_select';
+import { isMobileAgentName } from '../../../../common/agent_name';
+import { useApmServiceContext } from '../../../context/apm_service/use_apm_service_context';
+import { useBreakpoints } from '../../../hooks/use_breakpoints';
+import { ApmDatePicker } from '../date_picker/apm_date_picker';
+import { KueryBar } from '../kuery_bar';
+import { TimeComparison } from '../time_comparison';
+import { TransactionTypeSelect } from '../transaction_type_select';
 
 interface Props {
   hidden?: boolean;
